re-factor the code, improved the chapter generation logic
to fix:
1/. url - what happens when a user does not add a URL
we should by-pass this and generate a sub-domain
2/. if there is only one user in the chapter
what happens if a user creates a chapter, but they already are
an admin of another chapter and there is only one user there?
we don't want users to simply create chapters, ensure that the
a chapter always has an admin user.